refactor(server): isolate task store write paths
